+++ engine/runner/src/main/scala/org/enso/runner/ContextFactory.scala
@@ -23,6 +23,9 @@ class ContextFactory {
     * @param out the output stream for standard out
     * @param repl the Repl manager to use for this context
     * @param logLevel the log level for this context
+    * @param enableIrCaches whether or not IR caching should be enabled
+    * @param enableIrCacheReading whether or not IR cache reading should be
+    *                             enabled
     * @param strictErrors whether or not to use strict errors
     * @return configured Context instance
     */
@@ -33,6 +36,8 @@ class ContextFactory {
     repl: Repl,
     logLevel: LogLevel,
     logMasking: Boolean,
+    enableIrCaches: Boolean,
+    enableIrCacheReading: Boolean,
     strictErrors: Boolean = false
   ): PolyglotContext = {
     val context = Context
@@ -41,7 +46,12 @@ class ContextFactory {
       .allowAllAccess(true)
       .option(RuntimeOptions.PROJECT_ROOT, projectRoot)
       .option(RuntimeOptions.STRICT_ERRORS, strictErrors.toString)
-      .option(RuntimeOptions.DISABLE_IR_CACHES, "true")
+      .option(RuntimeOptions.WAIT_FOR_PENDING_SERIALIZATION_JOBS, "true")
+      .option(RuntimeOptions.DISABLE_IR_CACHES, (!enableIrCaches).toString)
+      .option(
+        RuntimeOptions.NO_READ_IR_CACHES,
+        (!enableIrCacheReading).toString
+      )
       .option(DebugServerInfo.ENABLE_OPTION, "true")
       .option(RuntimeOptions.LOG_MASKING, logMasking.toString)
       .option("js.foreign-object-prototype", "true")