+++ engine/runner/src/main/scala/org/enso/runner/Main.scala
@@ -1,6 +1,7 @@
 package org.enso.runner
 
 import akka.http.scaladsl.model.{IllegalUriException, Uri}
+import buildinfo.Info
 import cats.implicits._
 import com.typesafe.scalalogging.Logger
 import org.apache.commons.cli.{Option => CliOption, _}
@@ -43,6 +44,9 @@ object Main {
   private val IN_PROJECT_OPTION           = "in-project"
   private val VERSION_OPTION              = "version"
   private val JSON_OPTION                 = "json"
+  private val IR_CACHES_OPTION            = "ir-caches"
+  private val NO_IR_CACHES_OPTION         = "no-ir-caches"
+  private val NO_READ_IR_CACHES_OPTION    = "no-read-ir-caches"
   private val LOG_LEVEL                   = "log-level"
   private val LOGGER_CONNECT              = "logger-connect"
   private val NO_LOG_MASKING              = "no-log-masking"
@@ -51,6 +55,10 @@ object Main {
   private val AUTH_TOKEN                  = "auth-token"
 
   private lazy val logger = Logger[Main.type]
+
+  private def isDevBuild: Boolean = {
+    Info.ensoVersion.contains("SNAPSHOT")
+  }
 
   /** Builds the [[Options]] object representing the CLI syntax.
     *
@@ -229,6 +237,33 @@ object Main {
       .longOpt(AUTH_TOKEN)
       .desc("Authentication token for the upload.")
       .build()
+    val noReadIrCachesOption = CliOption.builder
+      .longOpt(NO_READ_IR_CACHES_OPTION)
+      .desc(
+        "Disables the reading of IR caches in the runtime if IR caching is enabled."
+      )
+      .build()
+
+    val irCachesOption = CliOption.builder
+      .longOpt(IR_CACHES_OPTION)
+      .desc(
+        "Enables IR caches. These are on by default in production builds " +
+        "and off by default in developer builds. You may not specify this " +
+        "option with `--no-ir-caches`."
+      )
+      .build()
+    val noIrCachesOption = CliOption.builder
+      .longOpt(NO_IR_CACHES_OPTION)
+      .desc(
+        "Disables IR caches. These are on by default in production builds " +
+        "and off by default in developer builds. You may not specify this " +
+        "option with `--ir-caches`."
+      )
+      .build()
+
+    val cacheOptionsGroup = new OptionGroup
+    cacheOptionsGroup.addOption(irCachesOption)
+    cacheOptionsGroup.addOption(noIrCachesOption)
 
     val options = new Options
     options
@@ -257,6 +292,8 @@ object Main {
       .addOption(uploadOption)
       .addOption(hideProgressOption)
       .addOption(authTokenOption)
+      .addOption(noReadIrCachesOption)
+      .addOptionGroup(cacheOptionsGroup)
 
     options
   }
@@ -286,11 +323,11 @@ object Main {
     * specifies the project name, otherwise the name is generated automatically.
     * The Enso version used in the project is set to the version of this runner.
     *
-    * @param path root path of the newly created project
-    * @param nameOption specifies the name of the created project
+    * @param path           root path of the newly created project
+    * @param nameOption     specifies the name of the created project
     * @param templateOption specifies the template of the created project
-    * @param authorName if set, sets the name of the author and maintainer
-    * @param authorEmail if set, sets the email of the author and maintainer
+    * @param authorName     if set, sets the name of the author and maintainer
+    * @param authorEmail    if set, sets the email of the author and maintainer
     */
   private def createNew(
     path: String,
@@ -337,17 +374,22 @@ object Main {
     * If `path` is a directory, so a project is run, a conflicting (pointing to
     * another project) `projectPath` should not be provided.
     *
-    * @param path path of the project or file to execute
-    * @param projectPath if specified, the script is run in context of a
-    *                    project located at that path
-    * @param logLevel log level to set for the engine runtime
-    * @param logMasking is the log masking enabled
+    * @param path           path of the project or file to execute
+    * @param projectPath    if specified, the script is run in context of a
+    *                       project located at that path
+    * @param logLevel       log level to set for the engine runtime
+    * @param logMasking     is the log masking enabled
+    * @param enableIrCaches are IR caches enabled
+    * @param readIrCaches whether IR caches should be read from if caching is
+    *                     enabled
     */
   private def run(
     path: String,
     projectPath: Option[String],
     logLevel: LogLevel,
-    logMasking: Boolean
+    logMasking: Boolean,
+    enableIrCaches: Boolean,
+    readIrCaches: Boolean
   ): Unit = {
     val file = new File(path)
     if (!file.exists) {
@@ -376,6 +418,8 @@ object Main {
       Repl(TerminalIO()),
       logLevel,
       logMasking,
+      enableIrCaches,
+      readIrCaches,
       strictErrors = true
     )
     if (projectMode) {
@@ -402,29 +446,41 @@ object Main {
     *
     * @param projectPath if specified, the docs is generated for a project
     *                    at the given path
-    * @param logLevel log level to set for the engine runtime
-    * @param logMasking is the log masking enabled
+    * @param logLevel    log level to set for the engine runtime
+    * @param logMasking  is the log masking enabled
+    * @param enableIrCaches are the IR caches enabled
+    * @param readIrCaches should the IR caches be read from if enabled
     */
   private def genDocs(
     projectPath: Option[String],
     logLevel: LogLevel,
-    logMasking: Boolean
+    logMasking: Boolean,
+    enableIrCaches: Boolean,
+    readIrCaches: Boolean
   ): Unit = {
     if (projectPath.isEmpty) {
       println("Path hasn't been provided.")
       exitFail()
     }
-    generateDocsFrom(projectPath.get, logLevel, logMasking)
+    generateDocsFrom(
+      projectPath.get,
+      logLevel,
+      logMasking,
+      enableIrCaches,
+      readIrCaches
+    )
     exitSuccess()
   }
 
   /** Subroutine of `genDocs` function.
     * Generates the documentation for given Enso project at given path.
     */
-  def generateDocsFrom(
+  private def generateDocsFrom(
     path: String,
     logLevel: LogLevel,
-    logMasking: Boolean
+    logMasking: Boolean,
+    enableIrCaches: Boolean,
+    readIrCaches: Boolean
   ): Unit = {
     val executionContext = new ContextFactory().create(
       path,
@@ -432,7 +488,9 @@ object Main {
       System.out,
       Repl(TerminalIO()),
       logLevel,
-      logMasking
+      logMasking,
+      enableIrCaches,
+      readIrCaches
     )
 
     val file = new File(path)
@@ -548,13 +606,17 @@ object Main {
     *
     * @param projectPath if specified, the REPL is run in context of a project
     *                    at the given path
-    * @param logLevel log level to set for the engine runtime
-    * @param logMasking is the log masking enabled
+    * @param logLevel    log level to set for the engine runtime
+    * @param logMasking  is the log masking enabled
+    * @param enableIrCaches are IR caches enabled
+    * @param readIrCaches should IR caches be read from if enabled
     */
   private def runRepl(
     projectPath: Option[String],
     logLevel: LogLevel,
-    logMasking: Boolean
+    logMasking: Boolean,
+    enableIrCaches: Boolean,
+    readIrCaches: Boolean
   ): Unit = {
     val mainMethodName = "internal_repl_entry_point___"
     val dummySourceToTriggerRepl =
@@ -571,7 +633,9 @@ object Main {
         System.out,
         Repl(TerminalIO()),
         logLevel,
-        logMasking
+        logMasking,
+        enableIrCaches,
+        readIrCaches
       )
     val mainModule =
       context.evalModule(dummySourceToTriggerRepl, replModuleName)
@@ -581,11 +645,11 @@ object Main {
 
   /** Handles `--server` CLI option
     *
-    * @param line a CLI line
+    * @param line     a CLI line
     * @param logLevel log level to set for the engine runtime
     */
   private def runLanguageServer(line: CommandLine, logLevel: LogLevel): Unit = {
-    val maybeConfig = parseSeverOptions(line)
+    val maybeConfig = parseServerOptions(line)
 
     maybeConfig match {
       case Left(errorMsg) =>
@@ -602,32 +666,32 @@ object Main {
     }
   }
 
-  private def parseSeverOptions(
+  private def parseServerOptions(
     line: CommandLine
   ): Either[String, LanguageServerConfig] =
     // format: off
     for {
-      rootId     <- Option(line.getOptionValue(ROOT_ID_OPTION))
-                      .toRight("Root id must be provided")
-                      .flatMap { id =>
-                        Either
-                          .catchNonFatal(UUID.fromString(id))
-                          .leftMap(_ => "Root must be UUID")
-                      }
-      rootPath   <- Option(line.getOptionValue(ROOT_PATH_OPTION))
-                      .toRight("Root path must be provided")
-      interface   = Option(line.getOptionValue(INTERFACE_OPTION))
-                      .getOrElse("127.0.0.1")
-      rpcPortStr  = Option(line.getOptionValue(RPC_PORT_OPTION)).getOrElse("8080")
-      rpcPort    <- Either
-                      .catchNonFatal(rpcPortStr.toInt)
-                      .leftMap(_ => "Port must be integer")
+      rootId <- Option(line.getOptionValue(ROOT_ID_OPTION))
+        .toRight("Root id must be provided")
+        .flatMap { id =>
+          Either
+            .catchNonFatal(UUID.fromString(id))
+            .leftMap(_ => "Root must be UUID")
+        }
+      rootPath <- Option(line.getOptionValue(ROOT_PATH_OPTION))
+        .toRight("Root path must be provided")
+      interface = Option(line.getOptionValue(INTERFACE_OPTION))
+        .getOrElse("127.0.0.1")
+      rpcPortStr = Option(line.getOptionValue(RPC_PORT_OPTION)).getOrElse("8080")
+      rpcPort <- Either
+        .catchNonFatal(rpcPortStr.toInt)
+        .leftMap(_ => "Port must be integer")
       dataPortStr = Option(line.getOptionValue(DATA_PORT_OPTION)).getOrElse("8081")
-      dataPort   <- Either
-                      .catchNonFatal(dataPortStr.toInt)
-                      .leftMap(_ => "Port must be integer")
+      dataPort <- Either
+        .catchNonFatal(dataPortStr.toInt)
+        .leftMap(_ => "Port must be integer")
     } yield boot.LanguageServerConfig(interface, rpcPort, dataPort, rootId, rootPath)
-    // format: on
+  // format: on
 
   /** Prints the version of the Enso executable.
     *
@@ -740,21 +804,27 @@ object Main {
         line.getOptionValue(RUN_OPTION),
         Option(line.getOptionValue(IN_PROJECT_OPTION)),
         logLevel,
-        logMasking
+        logMasking,
+        shouldEnableIrCaches(line),
+        shouldReadIrCaches(line)
       )
     }
     if (line.hasOption(REPL_OPTION)) {
       runRepl(
         Option(line.getOptionValue(IN_PROJECT_OPTION)),
         logLevel,
-        logMasking
+        logMasking,
+        shouldEnableIrCaches(line),
+        shouldReadIrCaches(line)
       )
     }
     if (line.hasOption(DOCS_OPTION)) {
       genDocs(
         Option(line.getOptionValue(IN_PROJECT_OPTION)),
         logLevel,
-        logMasking
+        logMasking,
+        shouldEnableIrCaches(line),
+        shouldReadIrCaches(line)
       )
     }
     if (line.hasOption(LANGUAGE_SERVER_OPTION)) {
@@ -764,5 +834,32 @@ object Main {
       printHelp(options)
       exitFail()
     }
+  }
+
+  /** Checks whether IR caching should be enabled.o
+    *
+    * The (mutually exclusive) flags can control it explicitly, otherwise it
+    * defaults to off in development builds and on in production builds.
+    *
+    * @param line the command-line
+    * @return `true` if caching should be enabled, `false`, otherwise
+    */
+  private def shouldEnableIrCaches(line: CommandLine): Boolean = {
+    if (line.hasOption(IR_CACHES_OPTION)) {
+      true
+    } else if (line.hasOption(NO_IR_CACHES_OPTION)) {
+      false
+    } else {
+      !isDevBuild
+    }
+  }
+
+  /** Determines whether IR caches should be read from if caching is enabled.
+    *
+    * @param line the command line
+    * @return `true` if caches should be read from, `false` otherwise
+    */
+  private def shouldReadIrCaches(line: CommandLine): Boolean = {
+    !line.hasOption(NO_READ_IR_CACHES_OPTION)
   }
 }