+++ engine/runtime/src/main/scala/org/enso/compiler/SerializationManager.scala
@@ -1,6 +1,7 @@
 package org.enso.compiler
 
 import com.oracle.truffle.api.TruffleLogger
+import com.oracle.truffle.api.source.Source
 import org.enso.compiler.core.IR
 import org.enso.interpreter.runtime.Module
 import org.enso.pkg.QualifiedName
@@ -66,6 +67,11 @@ class SerializationManager(compiler: Compiler) {
     * In addition, this method handles breaking links between modules contained
     * in the IR to ensure safe serialization.
     *
+    * It is responsible for taking a "snapshot" of the relevant module state at
+    * the point at which serialization is requested. This is due to the fact
+    * that serialization happens in a separate thread and the module may be
+    * mutated beneath it.
+    *
     * @param module the module to serialize
     * @return `true` if `module` has been scheduled for serialization, `false`
     *         otherwise
@@ -75,14 +81,18 @@ class SerializationManager(compiler: Compiler) {
       debugLogLevel,
       s"Requesting serialization for module [${module.getName}]."
     )
-    val duplicatedIr = module.getIr.duplicate()
+    val duplicatedIr = compiler.updateMetadata(
+      module.getIr,
+      module.getIr.duplicate(keepIdentifiers = true)
+    )
     duplicatedIr.preorder.foreach(_.passData.prepareForSerialization(compiler))
 
     val task = doSerialize(
       module.getCache,
       duplicatedIr,
       module.getCompilationStage,
-      module.getName
+      module.getName,
+      module.getSource
     )
     if (compiler.context.getEnvironment.isCreateThreadAllowed) {
       isWaitingForSerialization.put(module.getName, task)
@@ -92,6 +102,64 @@ class SerializationManager(compiler: Compiler) {
     }
 
     true
+  }
+
+  /** Deserializes the requested module from the cache if possible.
+    *
+    * If the requested module is currently being serialized it will wait for
+    * completion before loading. If the module is queued for serialization it
+    * will evict it and not load from the cache (this is usually indicative of a
+    * programming bug).
+    *
+    * @param module the module to deserialize from the cache.
+    * @return [[Some]] when deserialization was successful, with `true` for
+    *         relinking being successful and `false` otherwise. [[None]] if the
+    *         cache could not be deserialized.
+    */
+  def deserialize(module: Module): Option[Boolean] = {
+    if (isWaitingForSerialization(module)) {
+      abort(module)
+      None
+    } else {
+      while (isSerializing(module)) {
+        Thread.sleep(100)
+      }
+
+      module.getCache.load(compiler.context) match {
+        case Some(ModuleCache.CachedModule(ir, stage, _)) =>
+          val relinkedIrChecks =
+            ir.preorder.map(_.passData.restoreFromSerialization(this.compiler))
+          module.unsafeSetIr(ir)
+          module.unsafeSetCompilationStage(stage)
+          module.setLoadedFromCache(true)
+          logger.log(
+            debugLogLevel,
+            s"Restored IR from cache for module [${module.getName}] at stage [$stage]."
+          )
+
+          if (!relinkedIrChecks.contains(false)) {
+            module.setHasCrossModuleLinks(true)
+            logger.log(
+              debugLogLevel,
+              s"Restored links (early phase) in module [${module.getName}]."
+            )
+            Some(true)
+          } else {
+            logger.log(
+              debugLogLevel,
+              s"Could not restore links (early phase) in module [${module.getName}]."
+            )
+            module.setHasCrossModuleLinks(false)
+            Some(false)
+          }
+        case None =>
+          logger.log(
+            debugLogLevel,
+            s"Unable to load a cache for module [${module.getName}]."
+          )
+          None
+      }
+    }
   }
 
   /** Checks if the provided module is in the process of being serialized.
@@ -155,6 +223,10 @@ class SerializationManager(compiler: Compiler) {
       while (!pool.isTerminated) {
         pool.awaitTermination(500, TimeUnit.MILLISECONDS)
       }
+
+      pool.shutdownNow()
+      Thread.sleep(100)
+      logger.log(debugLogLevel, "Serialization manager has been shut down.")
     }
   }
 
@@ -172,13 +244,15 @@ class SerializationManager(compiler: Compiler) {
     * @param ir the IR for the module being serialized
     * @param stage the compilation stage of the module
     * @param name the name of the module being serialized
+    * @param source the source of the module being serialized
     * @return the task that serialies the provided `ir`
     */
   private def doSerialize(
     cache: ModuleCache,
     ir: IR.Module,
     stage: Module.CompilationStage,
-    name: QualifiedName
+    name: QualifiedName,
+    source: Source
   ): Runnable = { () =>
     startSerializing(name)
     logger.log(
@@ -190,7 +264,10 @@ class SerializationManager(compiler: Compiler) {
         if (stage.isAtLeast(Module.CompilationStage.AFTER_STATIC_PASSES)) {
           Module.CompilationStage.AFTER_STATIC_PASSES
         } else stage
-      cache.save(ModuleCache.CachedModule(ir, fixedStage), compiler.context)
+      cache.save(
+        ModuleCache.CachedModule(ir, fixedStage, source),
+        compiler.context
+      )
     } catch {
       case e: NotSerializableException =>
         logger.log(