+++ engine/runtime/src/main/scala/org/enso/compiler/pass/IRPass.scala
@@ -60,6 +60,24 @@ trait IRPass extends Serializable {
     ir: IR.Expression,
     inlineContext: InlineContext
   ): IR.Expression
+
+  /** Updates the metadata in a copy of the IR when updating that metadata
+    * requires global state.
+    *
+    * This is usually the case in the presence of structures that are shared
+    * throughout the IR, and need to maintain that sharing for correctness. This
+    * must be called with `copyOfIr` as the result of an `ir.duplicate` call.
+    *
+    * Additionally this method _must not_ alter the structure of the IR. It
+    * should only update its metadata.
+    *
+    * @param sourceIr the IR being copied
+    * @param copyOfIr a duplicate of `sourceIr`
+    * @tparam T the concrete [[IR]] type
+    * @return the result of updating metadata in `copyOfIr` globally using
+    *         information from `sourceIr`
+    */
+  def updateMetadataInDuplicate[T <: IR](sourceIr: T, copyOfIr: T): T
 }
 object IRPass {