+++ engine/runtime/src/main/scala/org/enso/compiler/pass/resolve/DocumentationComments.scala
@@ -9,6 +9,8 @@ import org.enso.compiler.exception.CompilerError
 import org.enso.compiler.pass.IRPass
 import org.enso.compiler.pass.desugar.{ComplexType, GenerateMethodBodies}
 
+import scala.annotation.unused
+
 /** Associates doc comments with the commented entities as metadata.
   *
   * If the first module definition is a documentation comment, it is treated as
@@ -56,6 +58,12 @@ case object DocumentationComments extends IRPass {
     ir: IR.Expression,
     inlineContext: InlineContext
   ): IR.Expression = resolveExpression(ir)
+
+  /** @inheritdoc */
+  override def updateMetadataInDuplicate[T <: IR](
+    @unused sourceIr: T,
+    copyOfIr: T
+  ): T = copyOfIr
 
   // === Pass Internals =======================================================