+++ engine/runtime/src/test/scala/org/enso/interpreter/test/instrument/RuntimeStdlibTest.scala
@@ -60,6 +60,7 @@ class RuntimeStdlibTest
         .option(RuntimeOptions.INTERPRETER_SEQUENTIAL_COMMAND_EXECUTION, "true")
         .option(RuntimeServerInfo.ENABLE_OPTION, "true")
         .option(RuntimeOptions.INTERACTIVE_MODE, "true")
+        .option(RuntimeOptions.DISABLE_IR_CACHES, "true")
         .out(out)
         .serverTransport(runtimeServerEmulator.makeServerTransport)
         .build()